+++ test/native_test/native_test.c
@@ -7,7 +7,7 @@
 #include <stdlib.h>
 #include <stdbool.h>
 
-bool Function1Bool(bool x) { return x; }
+bool Function1Bool(bool x) { return !x; }
 
 uint8_t Function1Uint8(uint8_t x) { return x + 42; }