+++ test/native_test/native_test.dart
@@ -51,8 +51,8 @@ void main() {
       }
     });
     test('bool', () {
-      expect(bindings.Function1Bool(0), 0);
-      expect(bindings.Function1Bool(1), 1);
+      expect(bindings.Function1Bool(true), false);
+      expect(bindings.Function1Bool(false), true);
     });
     test('uint8_t', () {
       expect(bindings.Function1Uint8(pow(2, 8).toInt()), 42);